requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29
# data science
numpy
pandas
scikit-learn
seaborn
statsmodels
# tests/linter
black
coverage
flake8
pytest
yapf
# API
gcsfs
google-cloud-storage
google-cloud-bigquery
mlflow
s3fs
# utilities
six>=1.14
joblib
memoized-property
termcolor
jupyter-contrib-core==0.4.0
jupyter-contrib-nbextensions==0.5.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.5.0
jupyter-resource-usage==0.6.3
jupyter-server==1.21.0
jupyter_client==7.4.3
jupyter_core==4.11.2
jupyterlab==3.4.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
jupyterlab_server==2.16.1
ipykernel==6.15.3
ipympl==0.9.2
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==7.7.2

# API
fastapi
pytz
uvicorn
